cli: store thread count as uint32_t, validate before narrowing
opt_threads is validated to the range 1..512, yet was stored as int64_t
and re-cast to uint64_t / std::size_t at every consumer (scanner.cc,
qgram.cc, algod1_network.cc, algod1_fastidious.cc). Store it as
std::uint32_t instead.

Range validation moves into a free function validate_threading(int64_t)
that checks the signed value returned by args_long() and returns the
narrowed result, so negative or oversized input (which would otherwise
wrap around when cast to an unsigned type) is rejected before the
narrowing. max_threads is int64_t to match the signed value being
checked. The error message is unchanged. The eight consumer casts are
dropped (uint32_t widens implicitly to size_t/uint64_t), and the log
format specifier becomes PRIu32.

test_options.sh: 470 PASS, 0 FAIL.

+++ src/utils/algod1_fastidious.cc
@@ -466,7 +466,7 @@ namespace {
     light_state.amplicon = data.sequence_count() - 1;
     {
       auto const light_tr = utils::make_unique<ThreadRunner>(
-          static_cast<std::size_t>(parameters.opt_threads),
+          parameters.opt_threads,
           [&parameters, &data, &ampinfo_v, &swarminfo_v, &hash_table, &bloom_a, &bloom_f, &light_state, &progress_light](uint64_t /*nth_thread*/) -> void {
             mark_light_thread(parameters, data, ampinfo_v, swarminfo_v, hash_table, bloom_a, bloom_f, light_state, progress_light);
           });
@@ -500,7 +500,7 @@ namespace {
     heavy_state.amplicon_count = amplicons_in_large_clusters;
     {
       auto const heavy_tr = utils::make_unique<ThreadRunner>(
-          static_cast<std::size_t>(parameters.opt_threads),
+          parameters.opt_threads,
           [&parameters, &data, &ampinfo_v, &swarminfo_v, &hash_table, &bloom_a, &bloom_f, &heavy_state, &graft_state, &progress_heavy](uint64_t /*nth_thread*/) -> void {
             check_heavy_thread(parameters, data, ampinfo_v, swarminfo_v, hash_table, bloom_a, bloom_f, heavy_state, graft_state, progress_heavy);
           });
